refactor: extract custom-form-total-cell to own file
